fix(frontend-backend): Add the modifications so that the roles work and add the relationship to the users, add styles for a better interface

+++ Quinto_Trimestre/TeddyShop - FrontEnd/src/pages/Roles/Roles.jsx
@@ -11,18 +11,19 @@ import {
   TableRow,
   Paper,
   IconButton,
-  Dialog,
   DialogTitle,
-  DialogContent,
   DialogActions,
   Snackbar,
   Alert,
   Box,
   Select,
+  Switch,
   MenuItem,
   FormControl,
+  FormControlLabel,
   TablePagination
 } from "@mui/material";
+import Swal from 'sweetalert2';
 import { Edit, Delete, ArrowUpward, ArrowDownward } from "@mui/icons-material";
 import '../PagesStyle.css';
 import { getApiUrl } from '../../utils/apiConfig'
@@ -35,11 +36,10 @@ const Roles = () => {
   const [role, setRole] = useState({ nombre: "", estado: true });
   const [isEditing, setIsEditing] = useState(false);
   const [currentId, setCurrentId] = useState(null);
-  const [openDeleteDialog, setOpenDeleteDialog] = useState(false);
   const [openSnackbar, setOpenSnackbar] = useState(false);
   const [snackbarMessage, setSnackbarMessage] = useState("");
   const [page, setPage] = useState(0);
-  const [rowsPerPage, setRowsPerPage] = useState(10); // Set maximum rows per page
+  const [rowsPerPage, setRowsPerPage] = useState(10);
   const [searchTerm, setSearchTerm] = useState("");
   const [sortBy, setSortBy] = useState("nombre");
   const [sortOrder, setSortOrder] = useState("asc");
@@ -49,16 +49,16 @@ const Roles = () => {
     return token;
   };
 
-  const token = getAuthToken();
+  useEffect(() => {
+    fetchRoles();
+  }, []);
 
   const fetchRoles = async () => {
     try {
-      const response = await fetch(`${apiUrl}/roles`, {
-      method: 'GET',
-      headers: {
-        'Authorization': `Bearer ${token}`,  // Añadir el token al encabezado
-      },
-    });
+      const response = await fetch(`${apiUrl}/roles`);
+      if (!response.ok) {
+        throw new Error('Error al obtener los roles');
+      } 
       const data = await response.json();
       setRoles(data);
       setFilteredRoles(data);
@@ -69,47 +69,84 @@ const Roles = () => {
     }
   };
 
-  useEffect(() => {
-    fetchRoles();
-  }, []);
-
-  const handleInputChange = (e) => {
-    setRole({ ...role, [e.target.name]: e.target.value });
-  };
-
-  const handleEstadoChange = (event) => {
-    setRole({ ...role, estado: event.target.value });
-  };
-
-  const handleSaveRole = async () => {
-    const url = isEditing ? `${apiUrl}/roles/${currentId}` : `${apiUrl}/roles`;
-    const method = isEditing ? 'PUT' : 'POST';
-    const token = getAuthToken();  // Obtener el token
-  
+  const crearRol = async () => {
+    console.log("Datos enviados:", role); 
     try {
-      const response = await fetch(url, {
-        method,
+      const response = await fetch(`${apiUrl}/roles`, {
+        method: 'POST',
         headers: {
           'Content-Type': 'application/json',
-          'Authorization': `Bearer ${token}`,  // Añadir el token al encabezado
         },
         body: JSON.stringify(role),
       });
   
-      if (response.ok) {
-        fetchRoles();
-        setRole({ nombre: '', estado: true });
-        setIsEditing(false);
-        setCurrentId(null);
-        setSnackbarMessage(isEditing ? "Rol actualizado" : "Rol creado");
-        setOpenSnackbar(true);
+      if (!response.ok) {
+        const errorData = await response.json();
+        throw new Error(errorData.message || 'Error al crear el rol');
       }
+  
+      fetchRoles();
+      setRole({ nombre: '', estado: true });
+      setSnackbarMessage("Rol creado correctamente");
+      setOpenSnackbar(true);
     } catch (error) {
-      console.error(isEditing ? 'Error updating role:' : 'Error creating role:', error);
-      setSnackbarMessage("Error al guardar el rol");
+      console.error('Error al crear el rol:', error);
+      setSnackbarMessage(error.message || "Error al guardar el rol");
       setOpenSnackbar(true);
     }
   };
+  
+  const actualizarRol = async () => {
+    if (!currentId) return;
+  
+    try {
+      const response = await fetch(`${apiUrl}/roles/${currentId}`, {
+        method: 'PUT',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(role),
+      });
+  
+      if (!response.ok) {
+        const errorData = await response.json();
+        throw new Error(errorData.message || 'Error al actualizar el rol');
+      }
+  
+      fetchRoles();
+      setRole({ nombre: '', estado: true });
+      setIsEditing(false);
+      setCurrentId(null);
+      setSnackbarMessage("Rol actualizado correctamente");
+      setOpenSnackbar(true);
+    } catch (error) {
+      console.error('Error al actualizar el rol:', error);
+      setSnackbarMessage(error.message || "Error al guardar el rol");
+      setOpenSnackbar(true);
+    }
+  };
+  const resetRoleForm = () => {
+    setRole({ nombre: "", estado: true });
+    setIsEditing(false);
+    setCurrentId(null);
+  };
+  
+  const handleSaveRole = () => {
+    isEditing ? actualizarRol() : crearRol();
+  };
+  
+  const handleInputChange = (e) => {
+    const { name, value } = e.target;
+  
+    setRole(prevRole => ({
+      ...prevRole,
+      [name]: name === "estado" ? Boolean(value) : value  
+    }));
+  };
+  
+    const handleEstadoChange = (event) => {
+      setRole({ ...role, estado: Boolean(event.target.checked) });
+    };
 
   const handleEditClick = (role) => {
     setRole({ nombre: role.nombre, estado: role.estado });
@@ -117,44 +154,38 @@ const Roles = () => {
     setCurrentId(role._id);
   };
 
-  const handleDeleteClick = (id) => {
-    setCurrentId(id);
-    setOpenDeleteDialog(true);
-  };
-
-  const handleDeleteRole = async () => {
-    const token = getAuthToken();  // Obtener el token
+  const EliminarRol = async (id) => {
+    Swal.fire({
+      title: "¿Estás seguro?",
+      text: "Esta acción no se puede deshacer",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Sí, eliminar",
+      cancelButtonText: "Cancelar"
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          const response = await fetch(`${apiUrl}/roles/${id}`, {
+            method: 'DELETE',
+          });
   
-    try {
-      const response = await fetch(`${apiUrl}/roles/${currentId}`, {
-        method: 'DELETE',
-        headers: {
-          'Authorization': `Bearer ${token}`,  // Añadir el token al encabezado
-        },
-      });
-  
-      if (response.ok) {
-        fetchRoles();
-        setOpenDeleteDialog(false);
-        setSnackbarMessage("Rol eliminado");
-        setOpenSnackbar(true);
+          if (response.ok) {
+            fetchRoles();
+            Swal.fire("Eliminado", "El rol ha sido eliminado correctamente.", "success");
+          } else {
+            Swal.fire("Error", "No se pudo eliminar el rol.", "error");
+          }
+        } catch (error) {
+          console.error('Error al eliminar el rol:', error);
+          Swal.fire("Error", "Ocurrió un problema al eliminar el rol.", "error");
+        }
       }
-    } catch (error) {
-      console.error('Error deleting role:', error);
-      setSnackbarMessage("Error al eliminar el rol");
-      setOpenSnackbar(true);
-    }
+    });
   };
-  
-
-  const handleCloseDeleteDialog = () => {
-    setOpenDeleteDialog(false);
-    setCurrentId(null);
-  };
-
   const handleCloseSnackbar = () => {
     setOpenSnackbar(false);
-    setSnackbarMessage("");
   };
 
   const handleChangePage = (event, newPage) => {
@@ -202,21 +233,33 @@ const Roles = () => {
           <h1>Gestión de Roles</h1>
           <form noValidate autoComplete="off">
             <TextField label="Nombre del Rol" name="nombre" value={role.nombre} onChange={handleInputChange} fullWidth margin="normal" required />
-            <FormControl fullWidth margin="normal">
-              <p>Estado</p>
-              <Select name="estado" value={role.estado} onChange={handleEstadoChange}>
-                <MenuItem value={true}>Activo</MenuItem>
-                <MenuItem value={false}>Inactivo</MenuItem>
-              </Select>
-            </FormControl>
-            <Button variant="contained" onClick={handleSaveRole} style={{ marginTop: 16 }}>{isEditing ? "Actualizar Rol" : "Crear Rol"}</Button>
-          </form>
+            <FormControlLabel
+            control={<Switch checked={role.estado} onChange={handleEstadoChange} />}
+            label={role.estado ? "Activo" : "Inactivo"}
+            />
+            <Box sx={{ mt: 1, display: "flex", justifyContent: "center", gap: 2  }}>
+          <Button 
+            variant="contained" 
+            color="primary" 
+            sx={{ borderRadius: "8px", padding: "8px 18px" }}
+            onClick={handleSaveRole}
+          >
+            {isEditing ? "Actualizar Rol" : "Crear Rol"}
+          </Button>
+          <Button 
+           variant="outlined" 
+           color="secondary" 
+           sx={{ borderRadius: "8px", padding: "8px 18px" }}
+           onClick={resetRoleForm}>
+            Cancelar
+          </Button>
+        </Box>
 
+          </form>
           <Box display="flex" justifyContent="space-between" alignItems="center" mt={4}>
             <h2>Roles</h2>
             <TextField label="Buscar por nombre" variant="outlined" size="small" value={searchTerm} onChange={handleSearchChange} style={{ width: 250 }} />
           </Box>
-
           <TableContainer component={Paper} style={{ marginTop: 20, maxHeight: 500, overflowY: "auto" }}>
             <Table stickyHeader>
               <TableHead>
@@ -241,9 +284,20 @@ const Roles = () => {
                   <TableRow key={r._id}>
                     <TableCell>{r.nombre}</TableCell>
                     <TableCell>{r.estado ? "Activo" : "Inactivo"}</TableCell>
+                    
                     <TableCell>
-                      <IconButton onClick={() => handleEditClick(r)}><Edit /></IconButton>
-                      <IconButton onClick={() => handleDeleteClick(r._id)}><Delete /></IconButton>
+                    <IconButton 
+                      onClick={() => handleEditClick(r)} 
+                      color="primary" 
+                    >
+                      <Edit />
+                    </IconButton>
+                    <IconButton 
+                      onClick={() => EliminarRol(r._id)} 
+                      sx={{ color: "red", "&:hover": { color: "darkred" } }}
+                    >
+                      <Delete />
+                    </IconButton>
                     </TableCell>
                   </TableRow>
                 ))}
@@ -262,20 +316,9 @@ const Roles = () => {
           />
         </Container>
       </Box>
-
-      {/* Snackbar for notifications */}
       <Snackbar open={openSnackbar} autoHideDuration={6000} onClose={handleCloseSnackbar}>
         <Alert onClose={handleCloseSnackbar} severity="success">{snackbarMessage}</Alert>
       </Snackbar>
-
-      {/* Delete confirmation dialog */}
-      <Dialog open={openDeleteDialog} onClose={handleCloseDeleteDialog}>
-        <DialogTitle>¿Estás seguro de que quieres eliminar este rol?</DialogTitle>
-        <DialogActions>
-          <Button onClick={handleCloseDeleteDialog} color="primary">Cancelar</Button>
-          <Button onClick={handleDeleteRole} color="secondary">Eliminar</Button>
-        </DialogActions>
-      </Dialog>
     </Box>
   );
 };
